refactor: rename network package connectId → personalId
Move entire package directory from org.commcare.connect.network.connectId to
org.commcare.connect.network.personalId. Updates all package declarations, imports,
and references across the codebase to reflect the new package structure.
