+++ app/src/org/commcare/android/database/connect/models/ConnectLinkedAppRecord.java
@@ -1,7 +1,7 @@
 package org.commcare.android.database.connect.models;
 
 import org.commcare.android.storage.framework.Persisted;
-import org.commcare.connect.network.connectId.SsoToken;
+import org.commcare.connect.network.personalId.SsoToken;
 import org.commcare.models.framework.Persisting;
 import org.commcare.modern.database.Table;
 import org.commcare.modern.models.MetaField;
