+++ app/src/org/commcare/connect/ConnectAppLauncher.kt
@@ -10,7 +10,7 @@ import org.commcare.activities.DataPullController.DataPullMode
 import org.commcare.activities.LoginMode
 import org.commcare.android.database.app.models.UserKeyRecord
 import org.commcare.connect.database.ConnectUserDatabaseUtil
-import org.commcare.connect.network.connectId.LoginInvalidatedException
+import org.commcare.connect.network.personalId.LoginInvalidatedException
 import org.commcare.google.services.analytics.FirebaseAnalyticsUtil
 import org.commcare.login.AppSeater
 import org.commcare.login.AuthSource
