+++ app/src/org/commcare/connect/PersonalIdManager.java
@@ -27,9 +27,9 @@ import org.commcare.connect.database.ConnectAppDatabaseUtil;
 import org.commcare.connect.database.ConnectDatabaseHelper;
 import org.commcare.connect.database.ConnectJobUtils;
 import org.commcare.connect.database.ConnectUserDatabaseUtil;
-import org.commcare.connect.network.connectId.ConnectSsoHelper;
-import org.commcare.connect.network.connectId.ConnectSsoSyncHelper;
-import org.commcare.connect.network.connectId.TokenExceptionHandler;
+import org.commcare.connect.network.personalId.ConnectSsoHelper;
+import org.commcare.connect.network.personalId.ConnectSsoSyncHelper;
+import org.commcare.connect.network.personalId.TokenExceptionHandler;
 import org.commcare.connect.workers.ConnectHeartbeatWorker;
 import org.commcare.connect.workers.ConnectReleaseTogglesWorker;
 import org.commcare.core.network.AuthInfo;
