+++ app/src/org/commcare/connect/network/personalId/PersonalIdApiClient.java
@@ -1,4 +1,4 @@
-package org.commcare.connect.network.connectId;
+package org.commcare.connect.network.personalId;
 
 import org.commcare.connect.network.base.BaseApiClient;
