+++ app/src/org/commcare/connect/network/personalId/PersonalIdApiHandler.java
@@ -1,4 +1,4 @@
-package org.commcare.connect.network.connectId;
+package org.commcare.connect.network.personalId;
 
 import android.app.Activity;
 import android.content.Context;
@@ -16,18 +16,18 @@ import org.commcare.connect.network.base.BaseApiCallback;
 import org.commcare.connect.network.base.BaseApiHandler;
 import org.commcare.connect.network.base.BaseApiResponseParser;
 import org.commcare.connect.network.connect.parser.ConnectReleaseTogglesParser;
-import org.commcare.connect.network.connectId.parser.AddOrVerifyNameParser;
-import org.commcare.connect.network.connectId.parser.CompleteProfileResponseParser;
-import org.commcare.connect.network.connectId.parser.ConfirmBackupCodeResponseParser;
-import org.commcare.connect.network.connectId.parser.ConnectTokenResponseParser;
-import org.commcare.connect.network.connectId.parser.LinkHqWorkerResponseParser;
-import org.commcare.connect.network.connectId.parser.PersonalIdApiResponseParser;
-import org.commcare.connect.network.connectId.parser.ReportIntegrityResponseParser;
-import org.commcare.connect.network.connectId.parser.RetrieveChannelEncryptionKeyResponseParser;
-import org.commcare.connect.network.connectId.parser.RetrieveHqTokenResponseParser;
-import org.commcare.connect.network.connectId.parser.RetrieveNotificationsResponseParser;
-import org.commcare.connect.network.connectId.parser.RetrieveWorkHistoryResponseParser;
-import org.commcare.connect.network.connectId.parser.StartConfigurationResponseParser;
+import org.commcare.connect.network.personalId.parser.AddOrVerifyNameParser;
+import org.commcare.connect.network.personalId.parser.CompleteProfileResponseParser;
+import org.commcare.connect.network.personalId.parser.ConfirmBackupCodeResponseParser;
+import org.commcare.connect.network.personalId.parser.ConnectTokenResponseParser;
+import org.commcare.connect.network.personalId.parser.LinkHqWorkerResponseParser;
+import org.commcare.connect.network.personalId.parser.PersonalIdApiResponseParser;
+import org.commcare.connect.network.personalId.parser.ReportIntegrityResponseParser;
+import org.commcare.connect.network.personalId.parser.RetrieveChannelEncryptionKeyResponseParser;
+import org.commcare.connect.network.personalId.parser.RetrieveHqTokenResponseParser;
+import org.commcare.connect.network.personalId.parser.RetrieveNotificationsResponseParser;
+import org.commcare.connect.network.personalId.parser.RetrieveWorkHistoryResponseParser;
+import org.commcare.connect.network.personalId.parser.StartConfigurationResponseParser;
 import org.commcare.interfaces.base.BaseConnectView;
 import org.commcare.util.LogTypes;
 import org.javarosa.core.io.StreamsUtil;
