+++ app/src/org/commcare/connect/network/personalId/TokenUnavailableException.kt
@@ -1,4 +1,4 @@
-package org.commcare.connect.network.connectId
+package org.commcare.connect.network.personalId
 
 import java.io.IOException
