+++ app/src/org/commcare/connect/network/personalId/parser/ConnectTokenResponseParser.kt
@@ -1,4 +1,4 @@
-package org.commcare.connect.network.connectId.parser
+package org.commcare.connect.network.personalId.parser
 
 import org.commcare.android.database.connect.models.ConnectUserRecord
 import org.commcare.connect.ConnectConstants.CONNECT_KEY_EXPIRES
