+++ app/src/org/commcare/connect/network/personalId/parser/NotificationParseResult.kt
@@ -1,4 +1,4 @@
-package org.commcare.connect.network.connectId.parser
+package org.commcare.connect.network.personalId.parser
 
 import org.commcare.android.database.connect.models.ConnectMessagingChannelRecord
 import org.commcare.android.database.connect.models.ConnectMessagingMessageRecord
