+++ app/src/org/commcare/connect/network/personalId/parser/ReportIntegrityResponseParser.java
@@ -1,4 +1,4 @@
-package org.commcare.connect.network.connectId.parser;
+package org.commcare.connect.network.personalId.parser;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
