+++ app/src/org/commcare/connect/network/personalId/parser/RetrieveHqTokenResponseParser.kt
@@ -1,4 +1,4 @@
-package org.commcare.connect.network.connectId.parser
+package org.commcare.connect.network.personalId.parser
 
 import android.content.Context
 import org.commcare.CommCareApplication
@@ -6,7 +6,7 @@ import org.commcare.connect.ConnectConstants.CONNECT_KEY_EXPIRES
 import org.commcare.connect.ConnectConstants.CONNECT_KEY_TOKEN
 import org.commcare.connect.database.ConnectDatabaseHelper
 import org.commcare.connect.network.base.BaseApiResponseParser
-import org.commcare.connect.network.connectId.SsoToken
+import org.commcare.connect.network.personalId.SsoToken
 import org.commcare.core.network.AuthInfo.TokenAuth
 import org.javarosa.core.io.StreamsUtil
 import org.json.JSONException
