+++ app/src/org/commcare/connect/network/personalId/parser/RetrieveWorkHistoryResponseParser.kt
@@ -1,4 +1,4 @@
-package org.commcare.connect.network.connectId.parser
+package org.commcare.connect.network.personalId.parser
 
 import android.content.Context
 import org.commcare.android.database.connect.models.PersonalIdWorkHistory
