+++ app/src/org/commcare/connect/workers/ConnectReleaseTogglesWorker.kt
@@ -12,7 +12,7 @@ import androidx.work.WorkManager
 import androidx.work.WorkerParameters
 import org.commcare.android.database.connect.models.ConnectReleaseToggleRecord
 import org.commcare.connect.PersonalIdManager
-import org.commcare.connect.network.connectId.PersonalIdApiHandler
+import org.commcare.connect.network.personalId.PersonalIdApiHandler
 import org.javarosa.core.services.Logger
 import java.util.concurrent.TimeUnit
 import kotlin.coroutines.resume
