+++ app/src/org/commcare/personalId/photo/PersonalIdPhotoUpdater.kt
@@ -11,7 +11,7 @@ import org.commcare.activities.camera.MicroImageActivity
 import org.commcare.connect.database.ConnectUserDatabaseUtil
 import org.commcare.connect.network.base.BaseApiHandler.PersonalIdOrConnectApiErrorCodes
 import org.commcare.connect.network.base.PersonalIdOrConnectApiErrorHandler
-import org.commcare.connect.network.connectId.PersonalIdApiHandler
+import org.commcare.connect.network.personalId.PersonalIdApiHandler
 import org.commcare.dalvik.R
 import org.commcare.utils.ConnectivityStatus
 import org.commcare.views.dialogs.StandardAlertDialog
