+++ app/src/org/commcare/personalId/profile/PersonalIdProfileEditFragment.kt
@@ -11,7 +11,7 @@ import androidx.navigation.fragment.findNavController
 import org.commcare.activities.CommCareActivity
 import org.commcare.connect.network.base.BaseApiHandler.PersonalIdOrConnectApiErrorCodes
 import org.commcare.connect.network.base.PersonalIdOrConnectApiErrorHandler
-import org.commcare.connect.network.connectId.PersonalIdApiHandler
+import org.commcare.connect.network.personalId.PersonalIdApiHandler
 import org.commcare.dalvik.R
 import org.commcare.dalvik.databinding.PersonalidProfileEditScreenBinding
 import org.commcare.fragments.personalId.EmailHelper
