+++ app/src/org/commcare/tasks/DataPullTask.java
@@ -13,8 +13,8 @@ import org.commcare.android.database.app.models.UserKeyRecord;
 import org.commcare.android.database.user.models.ACase;
 import org.commcare.cases.ledger.Ledger;
 import org.commcare.cases.util.InvalidCaseGraphException;
-import org.commcare.connect.network.connectId.LoginInvalidatedException;
-import org.commcare.connect.network.connectId.TokenUnavailableException;
+import org.commcare.connect.network.personalId.LoginInvalidatedException;
+import org.commcare.connect.network.personalId.TokenUnavailableException;
 import org.commcare.core.encryption.CryptUtil;
 import org.commcare.core.network.AuthenticationInterceptor;
 import org.commcare.core.network.CaptivePortalRedirectException;
