+++ app/unit-tests/src/org/commcare/connect/network/PersonalIdMockApiServer.kt
@@ -8,8 +8,8 @@ import okhttp3.mockwebserver.MockResponse
 import okhttp3.mockwebserver.MockWebServer
 import okhttp3.mockwebserver.RecordedRequest
 import org.commcare.connect.network.base.BaseApiClient
-import org.commcare.connect.network.connectId.PersonalIdApiClient
-import org.commcare.connect.network.connectId.PersonalIdApiService
+import org.commcare.connect.network.personalId.PersonalIdApiClient
+import org.commcare.connect.network.personalId.PersonalIdApiService
 import org.robolectric.shadows.ShadowLooper
 import java.util.concurrent.TimeUnit
